+++ include/clap/clap.h
@@ -69,4 +69,5 @@
 #include "ext/draft/tuning.h"
 #include "ext/draft/configurable-audio-ports.h"
 #include "ext/draft/extensible-audio-ports.h"
+#include "ext/draft/incremental-state.h"
 #include "ext/draft/undo.h"