+++ ChangeLog.md
@@ -11,6 +11,10 @@
 
 * [undo.h](include/clap/ext/draft/undo.h): undo support
 
+## Draft Extensions
+
+* [incremental-state.h](include/clap/ext/draft/incremental-state.h): incremental state
+
 # Changes in 1.1.9
 
 * [entry.h](include/clap/entry.h): clarify what the `plugin_path` is on macOS